CRM-17179 - CiviCRM Contact Ref - Contact Subtype handling fix
civicrm_contact table uses CRM_Core_DAO::VALUE_SEPARATOR for
storing multiple subtypes for one contact. civicrm_contact_ref
ignored this and simply used SQL IN() to test for matching. This
resulted in that a contact had multiple subtypes, it never matched.
So that's why I use separate WHERE clause to be able to do LIKE
operator and surround the value with the separator to avoid problems
with such types where one type is the substring of the another.