CRM-17468 - Drupal Views integration - add support for "is_primary" in address/phone/mail for sort and filter.
This also includes a fix for CRM-15994, fixing sort for country/state/county so that
they sort by name, not ID. This fix creates Views Relationships for Address, Phone,
and Email, in which the user can stipulate is_primary and other properties of an
address/phone/email. If these stipulations aren't relevant, the user can configure the
sort or filter without a relationship.  This relationship-based approach allows the user
to make those selections once in the relationship, rather than repeating the same
selections for each sort/filter (e.g., setting is_primary for each of country, state,
county, and city when sorting on all those fields)