CRM-17917 - Routing - Aggregate all subpaths
In my system with the current D8, activating Civi leads to a WSOD with the
message "The website encountered an unexpected error.  Please try again
later.The website encountered an unexpected error.  Please try again later."

There appears to be an issue with handling the `{one}/{two}/{three}/{four}`
trick for capturing subpaths. This uses a different trick for mapping
all subparts of the path to `$extra`.