Allow settings_location.php to be outside civicrm directory (call it civicrm_settings_location.php
We have had a long-standing issue with scripts that bypass our main
route provider needing to find the civicrm.settings.php file, notably
in cases where symlinks or other are present.

I have been looking at whether we could use our route provider for
kcfinder or use an alternative to kcfinder (no obvious candidate
is winning from what I can see).

Using our route provider I have
https://github.com/civicrm/civicrm-core/compare/master...eileenmcnaughton:kcfinder?expand=1#diff-bfe84caf9a6aad542cf0d5177d7105c0R12

However, there are 2 issues I've failed to solve
1) kcfinder adds it's own js scripts with a path relative to
it's own files - see the output in the html file that I included in the
commit.
2) I couldn't seem to get down to just kcfinder output with no drupal
wrapper

Failing to find a reasonably acheivable way to do the above I think
making the use of a file to identify the location of civicrm.settings.php is worth doing.

This patch simply extends the search for location_settings.php
to the folders above the civicrm folder (or a sites folder within them, this being our config)
 - allowing the civicrm_location_settings.php to be in the sites folder
this seems consistent with a drupal approach and means the
recommended CiviCRM practice of deleting the civicrm folder & replacing it does not break
kcfinder.

This is a low-traffic file so any extra file_exists calls seem unimportant.