(NFC) Update version in header
This is a simple administrative update to the headers. It was generated with the command:

```
rgrep '| CiviCRM version 4.7' CRM/ Civi ang api bin extern install/ settings/ templates -l \
 | xargs sed -i'' "s/| CiviCRM version 4.7/| CiviCRM version 5  /g"
```

The inclusion of `|` aimed to avoid matching any non-header text (e.g. inline docs that
mentioned the version incidentally). But then I did a looser search and for just

```
rgrep 'CiviCRM version 4.7'
````

and manually patched the remainder.

Note: I'm not really keen on doing this every month, so I relaxed the header
statement -- instead of `CiviCRM version 5.0`, it's just `CiviCRM version 5`.