Use the backend theme when upgrading
Related core PR:
https://github.com/civicrm/civicrm-core/pull/25961

This change was initially introduced in CiviCRM 4.2 to avoid
issues when upgrading, but we could not find any reason to
keep it that way today. Using the backend theme will make it
easier to provide consistent theming accross CMS-es and also
usually avoids some frontend CMS features to kick-in, such as Views.